Add files via upload
Due to a lack of time on my end, I had chatgpt walk through the code's functionality. It is still really, really difficult to understand, so I plan on adding more files to the repository and annotating this file myself to explain financial ratios/metrics and machine learning concepts better. Please stay tuned